+++ worker/include/RTC/SeqManager.hpp
@@ -51,6 +51,8 @@ namespace RTC
 		T base{ 0 };
 		T maxOutput{ 0 };
 		T maxInput{ 0 };
+		T maxDropped{ 0 };
+		T maxForwarded{ 0 };
 		std::set<T, SeqLowerThan> dropped;
 	};
 } // namespace RTC